+++ src/numbers_theory/great_common_divisor.cpp
@@ -1,16 +1,16 @@
-
+#include "types.hpp"
 
 // Implement your logic here.
-int gcdRecursive(int a, int b) {
+i64 gcdRecursive(i64 a, i64 b) {
     if (b==0) return a;
-    int r=a%b;
+    i64 r=a%b;
     return gcdRecursive(b,r);
 }
 
 
-int gcdIterativeEuclides(int a, int b) {
+i64 gcdIterativeEuclides(i64 a, i64 b) {
     while (b>0) {
-        int r=a%b;
+        i64 r=a%b;
         a=b;
         b=r;
     }
